added nav and userard
