+++ src/api/fake.api/user.api.js
@@ -150,6 +150,14 @@ const fetchAll = () =>
         }, 2000);
     });
 
+const getById = (id) =>
+    new Promise((resolve) => {
+        window.setTimeout(function () {
+            resolve(users.find((user) => user._id === id));
+        }, 1000);
+    });
+
 export default {
-    fetchAll
+    fetchAll,
+    getById
 };
