--- src/app.jsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-import Users from "./components/users";
-
-const App = () => {
-    return <Users />;
-};
-
-export default App;
