+++ src/components/layouts/login.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Login = () => {
+    return <h1>Login</h1>;
+};
+
+export default Login;
