+++ src/components/layouts/main.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Main = () => {
+    return <h1>Main page</h1>;
+};
+
+export default Main;
