+++ src/components/tableBody.jsx
@@ -1,17 +1,26 @@
 import React from "react";
 import PropTypes from "prop-types";
+import { Link } from "react-router-dom";
 import _ from "lodash";
 
 const TableBody = ({ data, columns }) => {
     const renderContent = (item, column) => {
-        if (columns[column].component) {
-            const component = columns[column].component;
-            if (typeof component === "function") {
-                return component(item);
+        if (column !== "name") {
+            if (columns[column].component) {
+                const component = columns[column].component;
+                if (typeof component === "function") {
+                    return component(item);
+                }
+                return component;
             }
-            return component;
+            return _.get(item, columns[column].path);
+        } else {
+            return (
+                <Link to={`/users/${item._id}`}>
+                    {_.get(item, columns[column].path)}
+                </Link>
+            );
         }
-        return _.get(item, columns[column].path);
     };
     return (
         <tbody>
